feat: Complete Day 4 of the Advent of Code
I completed the two puzzles for Day 4 for the Advent of Code. I was a
little late to the party but I'm completely caught up now. Looking
forward to the next one. I did add a Regex dependency so I could easily
parse some of the slightly more difficult input.